Consolidate seeds into one module
