+++ priv/repo/csv_seed_data/fantasy_players.csv
@@ -222,24 +222,6 @@ KD Pick #15,5,true
 KD Pick #16,5,true
 KD Pick #17,5,true
 KD Pick #18,5,true
-KD Horse #01,5,false
-KD Horse #02,5,false
-KD Horse #03,5,false
-KD Horse #04,5,false
-KD Horse #05,5,false
-KD Horse #06,5,false
-KD Horse #07,5,false
-KD Horse #08,5,false
-KD Horse #09,5,false
-KD Horse #10,5,false
-KD Horse #11,5,false
-KD Horse #12,5,false
-KD Horse #13,5,false
-KD Horse #14,5,false
-KD Horse #15,5,false
-KD Horse #16,5,false
-KD Horse #17,5,false
-KD Horse #18,5,false
 Arsenal,6,false
 Basel,6,false
 Atletico Madrid   ,6,false
@@ -413,22 +395,6 @@ LLWS Pick #13,11,true
 LLWS Pick #14,11,true
 LLWS Pick #15,11,true
 LLWS Pick #16,11,true
-LLWS Team #01,11,false
-LLWS Team #02,11,false
-LLWS Team #03,11,false
-LLWS Team #04,11,false
-LLWS Team #05,11,false
-LLWS Team #06,11,false
-LLWS Team #07,11,false
-LLWS Team #08,11,false
-LLWS Team #09,11,false
-LLWS Team #10,11,false
-LLWS Team #11,11,false
-LLWS Team #12,11,false
-LLWS Team #13,11,false
-LLWS Team #14,11,false
-LLWS Team #15,11,false
-LLWS Team #16,11,false
 Alexandr Dolgopolov ,12,false
 Andy Murray ,12,false
 Bernard Tomic,12,false
@@ -566,3 +532,37 @@ San Jose Earthquakes,15,false
 Seattle Sounders FC,15,false
 Sporting Kansas City,15,false
 Vancouver Whitecaps FC,15,false
+KD Horse #01,5,false
+KD Horse #02,5,false
+KD Horse #03,5,false
+KD Horse #04,5,false
+KD Horse #05,5,false
+KD Horse #06,5,false
+KD Horse #07,5,false
+KD Horse #08,5,false
+KD Horse #09,5,false
+KD Horse #10,5,false
+KD Horse #11,5,false
+KD Horse #12,5,false
+KD Horse #13,5,false
+KD Horse #14,5,false
+KD Horse #15,5,false
+KD Horse #16,5,false
+KD Horse #17,5,false
+KD Horse #18,5,false
+LLWS Team #01,11,false
+LLWS Team #02,11,false
+LLWS Team #03,11,false
+LLWS Team #04,11,false
+LLWS Team #05,11,false
+LLWS Team #06,11,false
+LLWS Team #07,11,false
+LLWS Team #08,11,false
+LLWS Team #09,11,false
+LLWS Team #10,11,false
+LLWS Team #11,11,false
+LLWS Team #12,11,false
+LLWS Team #13,11,false
+LLWS Team #14,11,false
+LLWS Team #15,11,false
+LLWS Team #16,11,false
